Move pricing service ownership to core
