+++ src/core/mod.rs
@@ -30,6 +30,7 @@ pub mod models;
 pub mod net; // Network validation and safety utilities
 pub mod observability; // Advanced observability and monitoring
 pub mod pricing; // Shared pricing data types
+pub mod pricing_service; // Runtime pricing service
 pub mod providers;
 pub mod rate_limiter; // Rate limiting system
 pub mod realtime; // Realtime WebSocket API
