+++ src/core/pricing_service/mod.rs
@@ -1,0 +1,20 @@
+//! Unified pricing service using LiteLLM pricing data format.
+//!
+//! This service loads pricing data from LiteLLM's JSON format and provides
+//! unified cost calculation for all AI providers.
+
+mod cache;
+mod events;
+mod loader;
+mod service;
+mod types;
+
+#[cfg(test)]
+mod tests;
+
+// Re-export public types
+pub use service::PricingService;
+pub use types::{
+    CostRange, CostResult, CostType, LiteLLMModelInfo, PricingEventType, PricingStatistics,
+    PricingUpdateEvent,
+};
