+++ src/services/pricing/mod.rs
@@ -1,20 +1,6 @@
-//! Unified pricing service using LiteLLM pricing data format
-//!
-//! This service loads pricing data from LiteLLM's JSON format and provides
-//! unified cost calculation for all AI providers
+//! Compatibility re-exports for the canonical core pricing service.
 
-mod cache;
-mod events;
-mod loader;
-mod service;
-mod types;
-
-#[cfg(test)]
-mod tests;
-
-// Re-export public types
-pub use service::PricingService;
-pub use types::{
-    CostRange, CostResult, CostType, LiteLLMModelInfo, PricingEventType, PricingStatistics,
-    PricingUpdateEvent,
+pub use crate::core::pricing_service::{
+    CostRange, CostResult, CostType, LiteLLMModelInfo, PricingEventType, PricingService,
+    PricingStatistics, PricingUpdateEvent,
 };
